EES-1646 Reorganise chart components into separate directory
