+++ src/explore-education-statistics-admin/src/pages/release/datablocks/components/chart/ChartAxisConfiguration.module.scss
