+++ src/explore-education-statistics-admin/src/pages/release/datablocks/components/chart/ChartLegendConfiguration.module.scss
